+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Fix RedisTasksRepository.get method
+
 ### 0.24.0 - 2020-04-13
 
  - Improve background task controller