+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## CHANGELOG
 
-### UNRELEASED
+### 0.25.0 - 2020-04-14
 
  - Refactor background task controller to build unique tasks ids when locking args